\name{phyexe-package}
\alias{phyexe-package}
\alias{phyexe}
\docType{package}
\title{
A collection of executables for phylogenetics
}
\description{
A collection of executables, including phyml, phytime, and RAxML.
}
\details{
\tabular{ll}{
Package: \tab phyexe\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2013-04-26\cr
License: \tab See individual packages for license information\cr
}

}
\author{
Maintainer: Simon D.W. Frost <sdf22@cam.ac.uk>
}
\references{
phyml, phytime: \href{http://code.google.com/p/phyml/}{http://code.google.com/p/phyml/}
RaXML: \href{https://github.com/stamatak/standard-RAxML}{https://github.com/stamatak/standard-RAxML}
}
\keyword{ package }
\seealso{
}
\examples{
}
